+++ routes/countries.js
@@ -132,6 +132,11 @@ router.post("/", function (req, res) {
   const NewName = req.body.countries_name;
   let db = database.GetDB();
 
+  if (!req.userAdmin){
+    res.status(403).json({ message: "You are not authorised to post!" });
+    return;
+  }
+
   if (!NewName) {
     res.status(400).json({ message: "products_name was null or empty"});
     return;
@@ -178,6 +183,11 @@ router.patch("/:id", function (req, res) {
   const id = req.params.id;
   // res.status(404).json({ message: "category does not exist" + NewName + " " + id });
 
+  if (!req.userAdmin){
+    res.status(403).json({ message: "You are not authorised to patch!"});
+    return;
+  }
+
   if (!NewName) {
     res.status(400).json({ message: "products_name was null or empty"});
     return;
@@ -212,6 +222,11 @@ router.patch("/:id", function (req, res) {
 router.delete("/:id", function (req, res) {
   const id = req.params.id;
   db = database.GetDB();
+
+  if (!req.userAdmin){
+    res.status(403).json({ message: "You are not authorised to delete!"});
+    return;
+  }
 
   db.run("DELETE FROM Countries WHERE countries_id = " + id + ";");
   res.status(200).json({ message: "Deleted!" });
